+++ example/lib/date_picker_bottom_sheet.dart
@@ -189,7 +189,6 @@ class _DatePickerBottomSheetState extends State<DatePickerBottomSheet> {
       pickerTheme: DateTimePickerTheme(
         showTitle: _showTitle,
         confirm: Text('custom Done', style: TextStyle(color: Colors.red)),
-        cancel: Text('custom cancel', style: TextStyle(color: Colors.cyan)),
       ),
       minDateTime: DateTime.parse(MIN_DATETIME),
       maxDateTime: DateTime.parse(MAX_DATETIME),