+++ lib/src/widget/date_picker_title_widget.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-import '../date_picker_theme.dart';
 import '../date_picker_constants.dart';
+import '../date_picker_theme.dart';
 import '../i18n/date_picker_i18n.dart';
 
 /// DatePicker's title widget.
@@ -32,18 +32,8 @@ class DatePickerTitleWidget extends StatelessWidget {
       child: Row(
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
         children: <Widget>[
-          Container(
-            height: pickerTheme.titleHeight,
-            child: FlatButton(
-                child: _renderCancelWidget(context),
-                onPressed: () => this.onCancel()),
-          ),
-          Container(
-            height: pickerTheme.titleHeight,
-            child: FlatButton(
-                child: _renderConfirmWidget(context),
-                onPressed: () => this.onConfirm()),
-          ),
+          _renderCancelWidget(context),
+          _renderConfirmWidget(context),
         ],
       ),
     );
@@ -51,26 +41,71 @@ class DatePickerTitleWidget extends StatelessWidget {
 
   /// render cancel button widget
   Widget _renderCancelWidget(BuildContext context) {
+    if (isCustomTitleWidget()) {
+      // has custom title button widget
+      if (pickerTheme.cancel == null) {
+        return Offstage();
+      }
+    }
+
     Widget cancelWidget = pickerTheme.cancel;
     if (cancelWidget == null) {
       TextStyle textStyle = pickerTheme.cancelTextStyle ??
           TextStyle(
-              color: Theme.of(context).unselectedWidgetColor, fontSize: 16.0);
-      cancelWidget =
-          Text(DatePickerI18n.getLocaleCancel(locale), style: textStyle);
+            color: Theme.of(context).unselectedWidgetColor,
+            fontSize: 16.0,
+          );
+      cancelWidget = Text(
+        DatePickerI18n.getLocaleCancel(locale),
+        style: textStyle,
+      );
     }
-    return cancelWidget;
+
+    return Container(
+      height: pickerTheme.titleHeight,
+      child: FlatButton(
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.zero),
+        child: cancelWidget,
+        onPressed: () => this.onCancel(),
+      ),
+    );
   }
 
   /// render confirm button widget
   Widget _renderConfirmWidget(BuildContext context) {
+    if (isCustomTitleWidget()) {
+      // has custom title button widget
+      if (pickerTheme.confirm == null) {
+        return Offstage();
+      }
+    }
+
     Widget confirmWidget = pickerTheme.confirm;
     if (confirmWidget == null) {
       TextStyle textStyle = pickerTheme.confirmTextStyle ??
-          TextStyle(color: Theme.of(context).primaryColor, fontSize: 16.0);
-      confirmWidget =
-          Text(DatePickerI18n.getLocaleDone(locale), style: textStyle);
+          TextStyle(
+            color: Theme.of(context).primaryColor,
+            fontSize: 16.0,
+          );
+      confirmWidget = Text(
+        DatePickerI18n.getLocaleDone(locale),
+        style: textStyle,
+      );
     }
-    return confirmWidget;
+
+    return Container(
+      height: pickerTheme.titleHeight,
+      child: FlatButton(
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.zero),
+        child: confirmWidget,
+        onPressed: () => this.onConfirm(),
+      ),
+    );
+  }
+
+  bool isCustomTitleWidget() {
+    return pickerTheme.cancel != null ||
+        pickerTheme.confirm != null ||
+        pickerTheme.title != null;
   }
 }